feat: add useScrollLockEffect
