+++ src/components/AppRoot/AppRoot.tsx
@@ -11,10 +11,8 @@ import { useIsomorphicLayoutEffect } from "../../lib/useIsomorphicLayoutEffect";
 import { classScopingMode } from "../../lib/classScopingMode";
 import { IconSettingsProvider } from "@vkontakte/icons";
 import {
-  elementScrollController,
-  globalScrollController,
-  ScrollContext,
-  ScrollContextInterface,
+  ElementScrollController,
+  GlobalScrollController,
 } from "./ScrollContext";
 import { noop } from "../../lib/utils";
 import { warnOnce } from "../../lib/warnOnce";
@@ -59,7 +57,7 @@ export const AppRoot = withAdaptivity<AppRootProps>(
     const [portalRoot, setPortalRoot] = React.useState<HTMLDivElement | null>(
       null
     );
-    const { window, document } = useDOM();
+    const { document } = useDOM();
     const insets = useInsets();
     const { appearance } = React.useContext(ConfigProviderContext);
 
@@ -170,12 +168,10 @@ export const AppRoot = withAdaptivity<AppRootProps>(
         document!.documentElement.style.removeProperty("color-scheme");
     }, [appearance]);
 
-    const scrollController = React.useMemo<ScrollContextInterface>(
+    const ScrollController = React.useMemo(
       () =>
-        scroll === "contain"
-          ? elementScrollController(rootRef)
-          : globalScrollController(window, document),
-      [document, scroll, window]
+        scroll === "contain" ? ElementScrollController : GlobalScrollController,
+      [scroll]
     );
 
     const content = (
@@ -188,14 +184,14 @@ export const AppRoot = withAdaptivity<AppRootProps>(
           mode,
         }}
       >
-        <ScrollContext.Provider value={scrollController}>
+        <ScrollController elRef={rootRef}>
           <IconSettingsProvider
             classPrefix="vkui"
             globalClasses={!noLegacyClasses}
           >
             {children}
           </IconSettingsProvider>
-        </ScrollContext.Provider>
+        </ScrollController>
       </AppRootContext.Provider>
     );
