+++ src/testing/utils.tsx
@@ -142,12 +142,19 @@ export const mockScrollContext = (
 ): [React.FC, jest.Mock] => {
   const getScroll = () => ({ x: 0, y: getY() });
   const scrollTo = jest.fn();
+  const isScrollLock = false;
   const enableScrollLock = jest.fn();
   const disableScrollLock = jest.fn();
   return [
     (props) => (
       <ScrollContext.Provider
-        value={{ getScroll, scrollTo, enableScrollLock, disableScrollLock }}
+        value={{
+          getScroll,
+          scrollTo,
+          isScrollLock,
+          enableScrollLock,
+          disableScrollLock,
+        }}
       >
         {props.children}
       </ScrollContext.Provider>
